Have created initial models and migrated the database
Have created initial models and migrated the database

+++ TequilasRestaurant/Models/Product.cs
@@ -7,7 +7,7 @@ namespace TequilasRestaurant.Models
         public int ProductId { get; set; }
         public string? Name { get; set; }
         public string? Description { get; set; }
-        public float Price { get; set; }
+        public Decimal Price { get; set; }
         public int Stock { get; set; }
         public int CategoryId {  get; set; }
         public Category? Category { get; set; } // A product belongs to a category
